# expression: implement vectorized evaluation for `builtinCurrentTime0ArgSig` and `builtinCurrentTime1ArgSig` 