--- a/expression/builtin_time_vec_test.go
+++ b/expression/builtin_time_vec_test.go
@@ -39,7 +39,9 @@ var vecBuiltinTimeCases = map[string][]vecExprBenchCase{
 	ast.DayOfWeek:   {},
 	ast.DayOfYear:   {},
 	ast.Day:         {},
-	ast.CurrentTime: {},
+	ast.CurrentTime: {
+		{retEvalType: types.ETDuration},
+	},
 	ast.CurrentDate: {
 		{retEvalType: types.ETDatetime},
 	},