<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->
### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for builtinCurrentTime0ArgSig, for #12176
implement vectorized evaluation for builtinCurrentTime1ArgSig, for #12176
### What is changed and how it works?
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args 'builtinCurrentTime0ArgSig'
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-4   	1000000000	         0.00765 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinCurrentTime0ArgSig-VecBuiltinFunc-4         	  518395	      2350 ns/op	     224 B/op	      11 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinCurrentTime0ArgSig-NonVecBuiltinFunc-4      	     637	   1828290 ns/op	  229430 B/op	   11264 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	2.672s
```
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args 'builtinCurrentTime1ArgSig'
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-4   	1000000000	         0.00787 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinCurrentTime1ArgSig-VecBuiltinFunc-4         	     841	   1379328 ns/op	  196686 B/op	   10241 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinCurrentTime1ArgSig-NonVecBuiltinFunc-4      	     632	   1901938 ns/op	  229433 B/op	   11264 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	2.776s
```
### Check List <!--REMOVE the items that are not applicable-->
Tests <!-- At least one of them must be included. -->
- Unit test